refactor: "CarManager 삭제 및 CarFactory 생성" - Car 생성 책임만을 갖는 팩토리 클래스 생성
