+++ src/test/java/study/step3/CarFactoryTest.java
@@ -1,0 +1,20 @@
+package study.step3;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+import study.step3.factory.CarFactory;
+
+public class CarFactoryTest {
+
+  @DisplayName("자동차 댓수 만큼 자동차를 생성한다.")
+  @ParameterizedTest
+  @ValueSource(ints = {1, 2, 3, 4, 5})
+  public void create_carArr_numberOfCars(int input) {
+    CarFactory carFactory = new CarFactory();
+
+    assertThat(carFactory.createCars(input).size()).isEqualTo(input);
+  }
+}
